WinSetup: remade into a tabbed control with 3 pages
