+++ Engine/CMakeLists.txt
@@ -453,6 +453,10 @@ target_sources(engine
     platform/windows/setup/windialog.h
     platform/windows/setup/winpagedialog.cpp
     platform/windows/setup/winpagedialog.h
+    platform/windows/setup/basicpagedialog.cpp
+    platform/windows/setup/basicpagedialog.h
+    platform/windows/setup/advancedpagedialog.cpp
+    platform/windows/setup/advancedpagedialog.h
     platform/windows/minidump.cpp
     platform/windows/win_ex_handling.cpp
