+++ Engine/resource/version.rc
@@ -114,10 +114,21 @@ IDI_ICON                ICON                    "game-1.ico"
 // Dialog
 //
 
-IDD_SETUP DIALOGEX 0, 0, 518, 316
+IDD_SETUP DIALOGEX 0, 0, 301, 303
 STYLE DS_SETFONT | DS_MODALFRAME | DS_CENTER | WS_POPUP | WS_CAPTION | WS_SYSMENU
 CAPTION "AGS Game Settings"
 FONT 8, "Tahoma", 400, 0, 0x0
+BEGIN
+    CTEXT           "Static",IDC_VERSION,7,266,287,9
+    DEFPUSHBUTTON   "S&ave",IDOK,7,282,56,14
+    PUSHBUTTON      "&Save and Run",IDOKRUN,67,282,56,14
+    PUSHBUTTON      "Cancel",IDCANCEL,127,282,56,14
+    CONTROL         "",IDC_TABPANEL,"SysTabControl32",0x0,7,7,287,253
+END
+
+IDD_PAGE_BASIC DIALOGEX 0, 0, 270, 198
+STYLE DS_SETFONT | WS_CHILD
+FONT 8, "Tahoma", 400, 0, 0xCC
 BEGIN
     GROUPBOX        "Graphics settings",IDC_GFXOPTIONS,7,7,247,148
     CONTROL         "Game resolution: XXXX x XXXX",IDC_RESOLUTION,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,15,19,161,10
@@ -137,39 +148,46 @@ BEGIN
     GROUPBOX        "Gameplay settings",IDC_STATIC,7,158,247,33
     CONTROL         "Game language:",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,15,172,56,10
     COMBOBOX        IDC_LANGUAGE,71,170,170,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    CTEXT           "Static",IDC_VERSION,7,207,247,9
-    DEFPUSHBUTTON   "S&ave",IDOK,7,224,56,14
-    PUSHBUTTON      "&Save and Run",IDOKRUN,67,224,56,14
-    PUSHBUTTON      "Cancel",IDCANCEL,127,224,56,14
-    PUSHBUTTON      "Ad&vanced  >>>",IDC_ADVANCED,188,224,67,14
-    GROUPBOX        "Graphics options",IDC_STATIC,264,7,247,91
-    CONTROL         "Vertical sync",IDC_VSYNC,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,271,20,135,9
-    CONTROL         "Smooth scaled sprites",IDC_ANTIALIAS,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,271,33,138,9
+END
+
+IDD_PAGE_ADVANCED DIALOGEX 0, 0, 278, 233
+STYLE DS_SETFONT | WS_CHILD
+FONT 8, "Tahoma", 400, 0, 0xCC
+BEGIN
+    GROUPBOX        "Graphics options",IDC_STATIC,7,7,247,90
+    CONTROL         "Vertical sync",IDC_VSYNC,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,15,20,135,9
+    CONTROL         "Smooth scaled sprites",IDC_ANTIALIAS,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,15,33,138,9
     CONTROL         "Render sprites at screen resolution",IDC_RENDERATSCREENRES,
-                    "Button",BS_AUTOCHECKBOX | BS_LEFT | BS_VCENTER | WS_TABSTOP,271,46,172,9
-    CONTROL         "Sprite cache (RAM):",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,271,63,78,10
-    COMBOBOX        IDC_SPRITECACHE,354,61,118,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    CONTROL         "Texture cache (VRAM):",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,271,80,78,10
-    COMBOBOX        IDC_TEXTURECACHE,354,78,118,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    GROUPBOX        "Sound options",IDC_STATIC,264,100,247,64
-    LTEXT           "Audio driver:",IDC_STATIC,271,115,46,9
-    COMBOBOX        IDC_DIGISOUND,320,113,178,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_GROUP | WS_TABSTOP
-    CONTROL         "Use voice pack if available",IDC_VOICEPACK,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,271,131,160,9
-    CONTROL         "Sound cache (RAM):",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,271,146,78,10
-    COMBOBOX        IDC_SOUNDCACHE,354,144,118,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
-    GROUPBOX        "Mouse options",IDC_STATIC,264,167,247,60
-    CONTROL         "Auto lock to window",IDC_MOUSE_AUTOLOCK,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,271,182,80,10
-    LTEXT           "Cursor speed: Default",IDC_MOUSESPEED_TEXT,271,196,176,8
-    CONTROL         "",IDC_MOUSESPEED,"msctls_trackbar32",WS_TABSTOP,271,207,226,14
-    CONTROL         "Custom game saves path",IDC_CUSTOMSAVEDIRCHECK,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,271,233,201,10,WS_EX_TRANSPARENT
-    GROUPBOX        "",IDC_STATIC,264,234,247,30
-    EDITTEXT        IDC_CUSTOMSAVEDIR,271,245,210,14,ES_AUTOHSCROLL | ES_READONLY
-    PUSHBUTTON      "...",IDC_CUSTOMSAVEDIRBTN,485,245,20,14
+                    "Button",BS_AUTOCHECKBOX | BS_LEFT | BS_VCENTER | WS_TABSTOP,15,46,172,9
+    CONTROL         "Sprite cache (RAM):",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,15,63,78,10
+    COMBOBOX        IDC_SPRITECACHE,97,61,118,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+    CONTROL         "Texture cache (VRAM):",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,15,80,78,10
+    COMBOBOX        IDC_TEXTURECACHE,97,77,118,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+    GROUPBOX        "Sound options",IDC_STATIC,7,99,247,64
+    LTEXT           "Audio driver:",IDC_STATIC,15,114,46,9
+    COMBOBOX        IDC_DIGISOUND,63,112,178,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_GROUP | WS_TABSTOP
+    CONTROL         "Use voice pack if available",IDC_VOICEPACK,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,15,130,160,9
+    CONTROL         "Sound cache (RAM):",IDC_STATIC,"Static",SS_LEFTNOWORDWRAP | WS_GROUP,15,145,78,10
+    COMBOBOX        IDC_SOUNDCACHE,97,144,118,100,CBS_DROPDOWNLIST | WS_VSCROLL | WS_TABSTOP
+    GROUPBOX        "Mouse options",IDC_STATIC,7,166,247,60
+    CONTROL         "Auto lock to window",IDC_MOUSE_AUTOLOCK,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,15,181,80,10
+    LTEXT           "Cursor speed: Default",IDC_MOUSESPEED_TEXT,15,195,176,8
+    CONTROL         "",IDC_MOUSESPEED,"msctls_trackbar32",WS_TABSTOP,15,206,226,14
+END
+
+IDD_PAGE_PATHS DIALOGEX 0, 0, 276, 88
+STYLE DS_SETFONT | WS_CHILD
+FONT 8, "Tahoma", 400, 0, 0xCC
+BEGIN
+    CONTROL         "Custom game saves path",IDC_CUSTOMSAVEDIRCHECK,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,15,7,201,10,WS_EX_TRANSPARENT
+    GROUPBOX        "",IDC_STATIC,7,8,247,30
+    EDITTEXT        IDC_CUSTOMSAVEDIR,15,19,210,14,ES_AUTOHSCROLL | ES_READONLY
+    PUSHBUTTON      "...",IDC_CUSTOMSAVEDIRBTN,229,19,20,14
     CONTROL         "Custom game shared data path",IDC_CUSTOMAPPDATADIRCHECK,
-                    "Button",BS_AUTOCHECKBOX | WS_TABSTOP,271,266,201,10,WS_EX_TRANSPARENT
-    GROUPBOX        "",IDC_STATIC,264,267,247,30
-    EDITTEXT        IDC_CUSTOMAPPDATADIR,271,278,210,14,ES_AUTOHSCROLL | ES_READONLY
-    PUSHBUTTON      "...",IDC_CUSTOMAPPDATADIRBTN,485,278,20,14
+                    "Button",BS_AUTOCHECKBOX | WS_TABSTOP,15,40,201,10,WS_EX_TRANSPARENT
+    GROUPBOX        "",IDC_STATIC,7,41,247,30
+    EDITTEXT        IDC_CUSTOMAPPDATADIR,15,52,210,14,ES_AUTOHSCROLL | ES_READONLY
+    PUSHBUTTON      "...",IDC_CUSTOMAPPDATADIRBTN,229,52,20,14
 END
 
 
@@ -184,11 +202,33 @@ BEGIN
     IDD_SETUP, DIALOG
     BEGIN
         LEFTMARGIN, 7
-        RIGHTMARGIN, 511
-        VERTGUIDE, 254
-        VERTGUIDE, 264
+        RIGHTMARGIN, 294
         TOPMARGIN, 7
-        BOTTOMMARGIN, 309
+        BOTTOMMARGIN, 296
+    END
+
+    IDD_PAGE_BASIC, DIALOG
+    BEGIN
+        LEFTMARGIN, 7
+        RIGHTMARGIN, 263
+        TOPMARGIN, 7
+        BOTTOMMARGIN, 191
+    END
+
+    IDD_PAGE_ADVANCED, DIALOG
+    BEGIN
+        LEFTMARGIN, 7
+        RIGHTMARGIN, 271
+        TOPMARGIN, 7
+        BOTTOMMARGIN, 226
+    END
+
+    IDD_PAGE_PATHS, DIALOG
+    BEGIN
+        LEFTMARGIN, 7
+        RIGHTMARGIN, 269
+        TOPMARGIN, 7
+        BOTTOMMARGIN, 81
     END
 END
 #endif    // APSTUDIO_INVOKED
@@ -200,6 +240,21 @@ END
 //
 
 IDD_SETUP AFX_DIALOG_LAYOUT
+BEGIN
+    0
+END
+
+IDD_PAGE_BASIC AFX_DIALOG_LAYOUT
+BEGIN
+    0
+END
+
+IDD_PAGE_ADVANCED AFX_DIALOG_LAYOUT
+BEGIN
+    0
+END
+
+IDD_PAGE_PATHS AFX_DIALOG_LAYOUT
 BEGIN
     0
 END
