+++ Solutions/Engine.App/Engine.App.vcxproj
@@ -652,6 +652,8 @@
     <ClCompile Include="..\..\Engine\platform\base\agsplatformdriver.cpp" />
     <ClCompile Include="..\..\Engine\platform\base\sys_main.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\acplwin.cpp" />
+    <ClCompile Include="..\..\Engine\platform\windows\setup\advancedpagedialog.cpp" />
+    <ClCompile Include="..\..\Engine\platform\windows\setup\basicpagedialog.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\setup\winapihelpers.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\setup\windialog.cpp" />
     <ClCompile Include="..\..\Engine\platform\windows\setup\winpagedialog.cpp" />
@@ -924,6 +926,8 @@
     <ClInclude Include="..\..\Engine\platform\base\agsplatformdriver.h" />
     <ClInclude Include="..\..\Engine\platform\base\sys_main.h" />
     <ClInclude Include="..\..\Engine\platform\windows\debug\namedpipesagsdebugger.h" />
+    <ClInclude Include="..\..\Engine\platform\windows\setup\advancedpagedialog.h" />
+    <ClInclude Include="..\..\Engine\platform\windows\setup\basicpagedialog.h" />
     <ClInclude Include="..\..\Engine\platform\windows\setup\winapihelpers.h" />
     <ClInclude Include="..\..\Engine\platform\windows\setup\windialog.h" />
     <ClInclude Include="..\..\Engine\platform\windows\setup\winpagedialog.h" />
